// src/auth/Types/roles.ts
export enum UserRoles {
    ADMIN = 'admin',
    PRODUCTION_MANAGER = 'pm',
    EDITOR = 'editor',
  DEPARTMENT_USER = 'user',
  STORE = 'store',
    SUPER_ADMIN = 'super_admin'
} 

export enum AllowedUserRoles {
  PRODUCTION_MANAGER = 'pm',
  EDITOR = 'editor',
  DEPARTMENT_USER = 'user',
  STORE = 'store',
  ADMIN='admin'
}